Fix a bug in MT include file
The include file for Oscilators for MT4 is missing the body code for CiATR::GetData. That causes a compile error.
This fix will avoid the error without changing the MT supplied code but if the bug is later fixed you will need to remove the function from the config file.